# Conflicts:
#	src/openMVG/graph/graph_graphviz_export.hpp
#	src/openMVG/matching_image_collection/Cascade_Hashing_Matcher_Regions_AllInMemory.cpp
#	src/openMVG/matching_image_collection/Cascade_Hashing_Matcher_Regions_AllInMemory.hpp
#	src/openMVG/matching_image_collection/Matcher.hpp
#	src/openMVG/matching_image_collection/Matcher_Regions_AllInMemory.cpp
#	src/openMVG/matching_image_collection/Matcher_Regions_AllInMemory.hpp
#	src/openMVG/sfm/pipelines/global/sfm_global_engine_relative_motions.cpp
#	src/openMVG/sfm/pipelines/sfm_features_provider.hpp
#	src/openMVG/sfm/pipelines/sfm_regions_provider.hpp
#	src/software/SfM/main_ComputeMatches.cpp

+++ src/openMVG/features/mser/mser_test.cpp
@@ -4,10 +4,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "testing/testing.h"
-#include "openMVG/image/image.hpp"
-#include "openMVG/features/mser/mser.hpp"
 #include "openMVG/features/feature.hpp"
+#include "openMVG/features/mser/mser.hpp"
+#include "openMVG/image/image.hpp"
+
+#include "testing/testing.h"
+
 
 using namespace openMVG ;
 using namespace image ; 
@@ -32,7 +34,7 @@ TEST( MSER , Extraction )
 
   // Export ellipse
   outimg = image ;
-  for( int i = 0 ; i < regs.size() ; ++i )
+  for( size_t i = 0 ; i < regs.size() ; ++i )
   {
     double ex , ey ;
     double Mx, My ;
