+++ src/openMVG/image/image_io_test.cpp
@@ -5,12 +5,13 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
+#include "openMVG/image/image.hpp"
+
+#include "testing/testing.h"
+
 #include <cstdio>
 #include <iostream>
 #include <string>
-
-#include "openMVG/image/image.hpp"
-#include "testing/testing.h"
 
 using namespace openMVG;
 using namespace openMVG::image;
@@ -226,7 +227,7 @@ TEST(ImageHeader, AllFormats) {
 
   const std::vector<std::string> ext_Type = {"jpg", "png", "tif", "png", "pgm"};
   const int image_border_size = 10;
-  for (int i=0; i < ext_Type.size(); ++i)
+  for (size_t i=0; i < ext_Type.size(); ++i)
   {
     std::ostringstream os;
     os << "img" << "." << ext_Type[i];
